refactor: move root graphql schema file to avoid circular imports
